♻️ Refactor: fetch함수에서 에러가 났을 경우를 고려하여 try-catch 도입 및 파일 이동
